St 1 centos rhel 71 issues

+++ test/cookbooks/fake/libraries/helper.rb
@@ -31,8 +31,12 @@ module EphemeralLvmTest
     def self.create_loop_devices(devices)
       Array(devices).each do |device|
         num = device.slice(/\d+/)
-        shell_out!("dd if=/dev/zero of=/vfile#{num} bs=1024 count=65536")
-        shell_out!("losetup #{device} /vfile#{num}")
+        if File.exists?("/vfile#{num}")
+          puts "/vfile#{num} previously created"
+        else
+          shell_out!("dd if=/dev/zero of=/vfile#{num} bs=1024 count=65536")
+          shell_out!("losetup #{device} /vfile#{num}")
+        end
       end
     end
